feat(FunctionState): Changes function state lock metrics to µs
Function state' lock holding time and lock congestion time are changed
from ms to µs(Ms).

+++ include/faabric/state/FunctionState.h
@@ -103,7 +103,7 @@ class FunctionState
     size_t sharedMemSize = 0;
     void* sharedMemory = nullptr;
 
-    long tempLockAquireTime = 0;
+    long long tempLockAquireTime = 0;
 
     // std::unordered_map<std::string, std::vector<uint8_t>> state;
     FunctionStateMetrics metrics;
